task/cookie/requirements.txt
fastapi>=0.120.0
uvicorn>=0.38.0
pycryptodome>=3.23.0
jinja2>=3.1.4
python-multipart>=0.0.20